Corrige la fuite de cartes en pioche à la mort, le classement en jetons au lieu d'unités monétaires, une course d'initialisation JS masquant le mode smartphone, et un bug d'affichage i18n - retours de test monnaie libre + smartphone (PDF "Retours - 20260919")
Bug le plus important, root cause de plusieurs symptômes signalés
("à partir du 5ème tour... la pioche semblait manquer de cartes faibles",
et en troc+smartphone "à sa naissance... il n'avait rien") :
GameService.recordEvent(), cas DEATH suivi par smartphone - les cartes
détenues par un joueur AU MOMENT DE SA MORT disparaissaient purement et
simplement du jeu (ni rendues à la pioche partagée, ni conservées), alors
que dealFreshHandForPlayer, appelée dans la foulée pour sa renaissance,
RETIRE 4 nouvelles cartes de cette même pioche. Chaque mort était donc une
fuite nette d'au moins 4 cartes hors circulation - un puits sans fond qui
épuise mécaniquement la pioche, d'autant plus vite qu'il y a de morts.
Corrigé en rendant les cartes du mourant à LEUR pioche de niveau avant de
piocher sa main de renaissance, exactement comme le fait déjà un carré pour
les 4 cartes qu'il défausse. Vérifié par un nouveau test (voir plus bas) :
sans ce correctif, le total de cartes en circulation perd exactement 4
cartes à CHAQUE mort ; avec, il reste rigoureusement constant sur 12 tours
simulés × 4 à 20 joueurs.

Remarque sur la règle de recyclage proposée dans le PDF ("sous 10 cartes
d'une valeur, inclure les cartes du carré à la pioche avant de repiocher") :
déjà implémentée depuis les correctifs du 05-06/09/2026
(checkAndCashInSquares rend TOUJOURS les 4 cartes défaussées à leur pioche
AVANT d'y repiocher les 4 de remplacement, sans condition de seuil - plus
fort que ce qui était demandé). La fuite corrigée ci-dessus était le vrai
goulot, pas la mécanique du carré elle-même.

Second bug d'investigation (remonté : "le joueur qui a vendu a conservé la
carte après la transaction" / "5 cartes identiques sans déclencher de
carré") : confirmé, par un nouveau test de 300 échanges aléatoires, que
computePlayerCardInventory (dérivé à 100% de l'historique Transaction/
CardSquareEvent, jamais un compteur mutable séparé) ne peut PAS dupliquer ou
perdre de carte, même sous forte charge. La vraie cause : les écrans
"Mes cartes"/"Classement" du joueur ne se rafraîchissaient QUE sur un
message WebSocket ciblé (vente confirmée avec la modal encore ouverte,
carré encaissé) - jamais par le rafraîchissement périodique déjà en place
pour le reste de l'écran (5 secondes). Un vendeur qui referme sa modal avant
que l'acheteur ne scanne (cas courant), ou un message WebSocket manqué lors
d'un accroc réseau, laissait ces deux écrans figés sur un instantané périmé
indéfiniment. Corrigé en les alignant sur le même rafraîchissement
périodique, uniquement quand ils sont l'écran actuellement affiché. Vérifié
en conditions réelles (Playwright) : une vente déclenchée depuis l'extérieur
pendant qu'un joueur reste simplement sur "Mes cartes" (jamais de clic) se
répercute bien sur son écran sous 7 secondes.

Classement en jetons au lieu d'unités monétaires (mockup fourni) :
GameService.computeLeaderboard utilisait computeTradeBalance (jetons
FAIBLES ÉQUIVALENTS) sans jamais le convertir par "Valeur d'une pièce
faible", contrairement au reste de l'écran joueur. Corrigé pour les joueurs
RÉELLEMENT suivis par smartphone uniquement (jamais la monnaie classique,
qui garde son calcul historique inchangé - principe "jamais l'une n'affecte
l'autre" de CLAUDE.md). LeaderboardEntryDto.value passe de int à double
(la conversion peut être fractionnaire, ex. weakCoinValue=0.4).

Bug d'affichage i18n ("Attention aux erreurs de label par moment", capture
montrant la clé brute affichée) - deux causes combinées :
1. Le libellé "Valeur d'une pièce faible"/version smartphone est réécrit en
   JS (textContent = t(clé)) dès que la liste des plugins arrive - une
   requête réseau indépendante de celle qui charge les traductions. Si
   cette dernière n'a pas fini, t() retombe sur la clé brute, et rien ne le
   corrigeait plus jamais ensuite. Corrigé en tenant aussi à jour l'attribut
   data-i18n (déjà présent par défaut) : la prochaine applyTranslations()
   de i18n.js (déclenchée dès que le fichier .po termine de charger, ou à
   chaque changement de langue) retrouve alors la bonne clé et corrige le
   libellé toute seule, quel que soit l'ordre d'arrivée des deux requêtes.
2. Bug plus grave découvert en creusant celui-ci : refreshAppSettings()
   (qui peuple mAppSettings.gameMode, lu par toute la logique d'affichage
   classique/smartphone de l'écran "Nouvelle partie") était lancée en
   "fire-and-forget" juste avant renderGamesList(), qui déclenche sa PROPRE
   requête réseau indépendante (liste des plugins) puis sélectionne aussitôt
   le premier système de monnaie. Une vraie course : si la liste des plugins
   répondait en premier (fréquent), tout l'écran se construisait avec
   mAppSettings.gameMode encore à sa valeur PLACEHOLDER (undefined) - donc
   TOUJOURS avec la disposition CLASSIQUE, quel que soit le réglage réel,
   jusqu'au premier clic manuel de l'animateur sur un système de monnaie.
   Invisible tant que "classique" était le réglage par défaut de toute
   façon - mais affichait à tort l'écran classique dès la toute première
   ouverture de l'application maintenant que "smartphone" l'est devenu
   (voir le commit précédent, AppSettings.java). Corrigé en attendant
   explicitement la fin de refreshAppSettings() avant de lancer
   renderGamesList(). Vérifié par Playwright avec des délais réseau
   artificiels sur /api/settings, /api/plugins et les fichiers .po, seuls
   ou combinés : la disposition des champs reste désormais toujours
   correcte, et un éventuel flash de clé brute (uniquement si les DEUX
   réglages ET traductions sont anormalement lents en même temps) se
   corrige tout seul dès que les traductions arrivent.

Non traité dans ce commit, à discuter avec l'utilisateur : le report d'écart
d'arrondi entre le total des soldes joueurs et la masse monétaire globale,
et la proposition de le reporter sur la prochaine naissance. En mode strict
TRM (celui testé ici), la masse monétaire est TOUJOURS recalculée comme la
somme exacte des jetons réels à chaque tour/mort (voir
computeMoneyMassFromActivePlayersJetons) - les deux ne peuvent donc pas
diverger par construction, ce qui rend la mécanique proposée inapplicable
telle quelle dans ce mode. Le symptôme décrit correspond plutôt au mode NON
strict TRM (masse convergeant vers une cible indépendante du solde réel des
joueurs) - une clarification est nécessaire avant d'implémenter quoi que ce
soit ici pour ne pas se tromper de mécanisme.

Tests : nouveau GameServicePileConservationTest (geco-server, JPA direct sur
H2 en mémoire, jamais via HTTP) - 7 tests : conservation du nombre total de
cartes sur 80 morts successives, sur 12 tours simulés × {4, 8, 12, 20}
joueurs (toute la plage demandée), absence de duplication/perte sur 300
échanges aléatoires, et conversion correcte du classement en unités
monétaires. Chaque test de conservation a été vérifié comme un VRAI test de
régression : en annulant temporairement le correctif, testManyDeathsNever...
échoue dès la première mort (perd exactement 4 cartes), confirmant qu'il
capture bien le bug réel plutôt que de passer par construction.
mvn clean test : 26/26 tests verts (19 existants + 7 nouveaux). Vérifié
aussi en conditions réelles (./run.sh --rebuild + Playwright + appels API
directs) : partie libre+smartphone à weakCoinValue=0.4 créée, classement
affichant 7,2 (au lieu de 18 jetons bruts), rafraîchissement automatique de
"Mes cartes" confirmé sous 7 secondes sans navigation manuelle.

+++ geco-server/src/main/java/jyt/geconomicus/helper/server/Dtos.java
@@ -398,7 +398,17 @@ public class Dtos
 	}
 
 	// Réponse de GET /api/games/{id}/leaderboard - voir GameService.computeLeaderboard.
-	public record LeaderboardEntryDto(Integer playerId, String playerName, int value, int rank)
+	// value : double depuis le 19/09/2026 (remonté par l'utilisateur, PDF
+	// "Retours - 20260919" : "mettre l'affichage du classement en unités
+	// monétaires (il est actuellement en jetons)") - en monnaie libre/dette,
+	// ce classement affichait auparavant le nombre BRUT de jetons
+	// (GameService.computeTradeBalance, une valeur en jetons faibles
+	// ÉQUIVALENTS, jamais convertie par "Valeur d'une pièce faible"), alors
+	// que tout le reste de l'écran joueur (solde, dernière activité...)
+	// affiche systématiquement des UNITÉS MONÉTAIRES (voir jetonsToMoney
+	// côté client). int -> double : cette conversion peut être fractionnaire
+	// dès que weakCoinValue != 1 (ex. 7 jetons à weakCoinValue=0.4 = 2.8).
+	public record LeaderboardEntryDto(Integer playerId, String playerName, double value, int rank)
 	{
 	}
